Cast away a warning.
Alas, QByteArray's constructor takes an int, not a size_t, as a size.
Explicitly cast the size down to an int.

Reviewed-on: https://code.wireshark.org/review/14836

+++ ui/qt/follow_stream_dialog.cpp
@@ -461,7 +461,7 @@ FollowStreamDialog::readSslStream()
             size_t nchars = rec->data.data_len;
             // We want a deep copy.
             buffer.clear();
-            buffer.append((const char *) rec->data.data, nchars);
+            buffer.append((const char *) rec->data.data, (int)nchars);
 
             frs_return = showBuffer(buffer.data(), nchars,
                                      rec->is_from_server, rec->packet_num, global_pos);
